[수정] 관리자 페이지 매니저 관리 축제권한 기능 추가 및 유효성 검사 기능 수정

+++ back/KHFinalProject/src/main/java/com/zeus/admin/domain/AdminEventDTO.java
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 public class AdminEventDTO {
-    private int no;
+    private Long no;
     private String publicDataEventNo;
     private int userAccountNo;
     private String price;
